requirements.txt
lxml==6.0.0
pillow==11.3.0
python-docx==1.2.0
simple-term-menu==1.6.6
typing_extensions==4.14.1
